refactor: migrate Flink adapter Utils.applyTableChange and HoodieCatalogUtil to HoodieSchema
Flips the per-Flink-version DDL apply path off InternalSchemaChangeApplier:

  - Each adapter Utils.java (1.17.x .. 2.1.x) carries oldSchema and the
    accumulated newSchema as HoodieSchema, instantiating
    HoodieSchemaChangeApplier per change. The convertFunc parameter is
    now Function<LogicalType, HoodieSchema> so callers no longer need to
    bridge through InternalSchemaConverter.convertToField.
    parseColumnPosition returns Pair<ColumnPositionType, String> using
    the HoodieSchema-side enum from common.schema.evolution.

  - HoodieCatalogUtil.alterTable drops the bridge calls at the
    Utils.applyTableChange boundary; convertFunc collapses to the direct
    method reference HoodieSchemaConverter::convertToSchema and the
    HoodieFlinkWriteClient round-trip stays HoodieSchema-shaped end to
    end via getEvolutionSchemaAndMetaClient + commitTableChange(HoodieSchema, _).

InternalSchema / InternalSchemaChangeApplier / InternalSchemaConverter /
HoodieSchemaInternalSchemaBridge / Type imports drop from all 7 files.
The Flink DDL chain is now fully off internal.schema.
