+++ hudi-flink-datasource/hudi-flink2.1.x/src/main/java/org/apache/hudi/adapter/Utils.java
@@ -18,11 +18,11 @@
 
 package org.apache.hudi.adapter;
 
+import org.apache.hudi.common.schema.HoodieSchema;
+import org.apache.hudi.common.schema.evolution.ColumnPositionType;
+import org.apache.hudi.common.schema.evolution.HoodieSchemaChangeApplier;
 import org.apache.hudi.common.util.collection.Pair;
 import org.apache.hudi.exception.HoodieNotSupportedException;
-import org.apache.hudi.internal.schema.InternalSchema;
-import org.apache.hudi.common.schema.types.Type;
-import org.apache.hudi.internal.schema.action.InternalSchemaChangeApplier;
 
 import org.apache.flink.configuration.Configuration;
 import org.apache.flink.configuration.ReadableConfig;
@@ -49,9 +49,9 @@ import java.util.Collections;
 import java.util.List;
 import java.util.function.Function;
 
-import static org.apache.hudi.internal.schema.action.TableChange.ColumnPositionChange.ColumnPositionType.AFTER;
-import static org.apache.hudi.internal.schema.action.TableChange.ColumnPositionChange.ColumnPositionType.FIRST;
-import static org.apache.hudi.internal.schema.action.TableChange.ColumnPositionChange.ColumnPositionType.NO_OPERATION;
+import static org.apache.hudi.common.schema.evolution.ColumnPositionType.AFTER;
+import static org.apache.hudi.common.schema.evolution.ColumnPositionType.FIRST;
+import static org.apache.hudi.common.schema.evolution.ColumnPositionType.NO_OPERATION;
 
 /**
  * Adapter utils.
@@ -84,8 +84,8 @@ public class Utils {
         conf.get(ExecutionConfigOptions.TABLE_EXEC_SORT_ASYNC_MERGE_ENABLED));
   }
 
-  public static InternalSchema applyTableChange(InternalSchema oldSchema, List changes, Function<LogicalType, Type> convertFunc) {
-    InternalSchema newSchema = oldSchema;
+  public static HoodieSchema applyTableChange(HoodieSchema oldSchema, List changes, Function<LogicalType, HoodieSchema> convertFunc) {
+    HoodieSchema newSchema = oldSchema;
     for (Object change : changes) {
       TableChange tableChange = (TableChange) change;
       newSchema = applyTableChange(newSchema, tableChange, convertFunc);
@@ -116,16 +116,16 @@ public class Utils {
                     environment.getUserCodeClassLoader().asClassLoader()));
   }
 
-  private static InternalSchema applyTableChange(InternalSchema oldSchema, TableChange change, Function<LogicalType, Type> convertFunc) {
-    InternalSchemaChangeApplier changeApplier = new InternalSchemaChangeApplier(oldSchema);
+  private static HoodieSchema applyTableChange(HoodieSchema oldSchema, TableChange change, Function<LogicalType, HoodieSchema> convertFunc) {
+    HoodieSchemaChangeApplier changeApplier = new HoodieSchemaChangeApplier(oldSchema);
     if (change instanceof TableChange.AddColumn) {
       if (((TableChange.AddColumn) change).getColumn().isPhysical()) {
         TableChange.AddColumn add = (TableChange.AddColumn) change;
         Column column = add.getColumn();
         String colName = column.getName();
-        Type colType = convertFunc.apply(column.getDataType().getLogicalType());
+        HoodieSchema colType = convertFunc.apply(column.getDataType().getLogicalType());
         String comment = column.getComment().orElse(null);
-        Pair<org.apache.hudi.internal.schema.action.TableChange.ColumnPositionChange.ColumnPositionType, String> colPositionPair =
+        Pair<ColumnPositionType, String> colPositionPair =
             parseColumnPosition(add.getPosition());
         return changeApplier.applyAddChange(
             colName, colType, comment, colPositionPair.getRight(), colPositionPair.getLeft());
@@ -143,12 +143,12 @@ public class Utils {
     } else if (change instanceof TableChange.ModifyPhysicalColumnType) {
       TableChange.ModifyPhysicalColumnType modify = (TableChange.ModifyPhysicalColumnType) change;
       String colName = modify.getOldColumn().getName();
-      Type newColType = convertFunc.apply(modify.getNewType().getLogicalType());
+      HoodieSchema newColType = convertFunc.apply(modify.getNewType().getLogicalType());
       return changeApplier.applyColumnTypeChange(colName, newColType);
     } else if (change instanceof TableChange.ModifyColumnPosition) {
       TableChange.ModifyColumnPosition modify = (TableChange.ModifyColumnPosition) change;
       String colName = modify.getOldColumn().getName();
-      Pair<org.apache.hudi.internal.schema.action.TableChange.ColumnPositionChange.ColumnPositionType, String> colPositionPair =
+      Pair<ColumnPositionType, String> colPositionPair =
           parseColumnPosition(modify.getNewPosition());
       return changeApplier.applyReOrderColPositionChange(
           colName, colPositionPair.getRight(), colPositionPair.getLeft());
@@ -164,8 +164,8 @@ public class Utils {
     }
   }
 
-  private static Pair<org.apache.hudi.internal.schema.action.TableChange.ColumnPositionChange.ColumnPositionType, String> parseColumnPosition(TableChange.ColumnPosition colPosition) {
-    org.apache.hudi.internal.schema.action.TableChange.ColumnPositionChange.ColumnPositionType positionType;
+  private static Pair<ColumnPositionType, String> parseColumnPosition(TableChange.ColumnPosition colPosition) {
+    ColumnPositionType positionType;
     String position = "";
     if (colPosition instanceof TableChange.First) {
       positionType = FIRST;
